+++ CHANGELOG.md
@@ -4,8 +4,8 @@ SwiftyRSA Changelog
 # [master]
 
  - Fixed compilation warnings for Xcode 9.1 / 9.2.
-
- - Added `generateRSAKeyPair` to SwiftyRSA enum. [#106](https://github.com/TakeScoop/SwiftyRSA/issues/106)
+ - Added ability to generate a RSA key pair by using `SwiftyRSA.generateRSAKeyPair`.
+   [#106](https://github.com/TakeScoop/SwiftyRSA/issues/106)
 
 # [1.3.0]
 
@@ -107,7 +107,6 @@ We recommend to check out the new [usage instructions](./README.md) to migrate c
 Initial release.
 
 [master]: https://github.com/TakeScoop/SwiftyRSA/tree/master
-[1.4.0]: https://github.com/TakeScoop/SwiftyRSA/releases/tag/1.4.0
 [1.3.0]: https://github.com/TakeScoop/SwiftyRSA/releases/tag/1.3.0
 [1.2.0]: https://github.com/TakeScoop/SwiftyRSA/releases/tag/1.2.0
 [1.1.1]: https://github.com/TakeScoop/SwiftyRSA/releases/tag/1.1.1